// src/site-data/deserialize/validate/serialized-site-data-schema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SerializedSiteData {
  projects: SerializedProject[]
  technologies: SerializedTechnology[]
}
export interface SerializedProject {
  abrv?: string
  name: string
  urlSafeName?: string
  short: string
  longMdUrl?: string
  gitUrl?: string
  imgUrl?: string
  siteUrl?: string
  technologies: string[]
}
export interface SerializedTechnology {
  name: string
  urlSafeName?: string
  iconName?: string
  short?: string
  longMdUrl?: string
  projects?: string[]
}